config/requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
requests==2.27.1
django-crispy-forms==1.12.0
django-ckeditor==6.1.0
Pillow==8.3.1
django-import-export==2.7.1
ua-parser==0.10.0
user-agents==2.2.0
django-user-agents==0.4.0
python-memcached==1.59
six==1.16.0
channels==3.0.4
channels-redis==3.3.0
daphne==3.0.2
gunicorn==20.1.0
django-model-utils==4.1.1
